Object Polling, Client Map Init Flow, etc
gameplayManager, playerComponent etc... about game playing code, prefab system(not tested)

+++ Session/netcode/SessionNetworkDto.h
@@ -27,7 +27,7 @@ NLOHMANN_JSON_SERIALIZE_ENUM(GameMode,{
 struct player_dto {
     std::string id;
     std::string name;
-    uint8_t team;
+    int team;
     int kill = 0;
     int death = 0;
 };
